Brought in a dependency for alternative to window confirmation that I will implement at a later time. Added a window.confirm when deleting a user.

+++ frontend/src/screens/UserListScreen.js
@@ -27,7 +27,9 @@ const UserListScreen = ({ history }) => {
   }, [dispatch, history, successDelete, userInfo]);
 
   const deleteHandler = (id) => {
-    dispatch(deleteUser(id));
+    if (window.confirm("Are you sure?")) {
+      dispatch(deleteUser(id));
+    }
   };
 
   return (
